refactor: 관심사의 분리 위한 구조 개선(ResultsSection.tsx)
